requirements.txt
certifi==2019.6.16
pika==1.1.0
PyYAML==5.1.2